Reorganize project in modules
