+++ src/app/app.module.ts
@@ -1,48 +1,23 @@
-import {LOCALE_ID, NgModule} from '@angular/core';
+import {NgModule} from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
 import { AppComponent } from './app.component';
-import { FaceSnapComponent } from './face-snap/face-snap.component';
-import {registerLocaleData} from "@angular/common";
-import * as fr from '@angular/common/locales/fr';
-import { LightboxModule } from 'ngx-lightbox';
-
-import { FaceSnapListComponent } from './face-snap-list/face-snap-list.component';
-import { HeaderComponent } from './header/header.component';
 import {AppRoutingModule} from "./app-routing.module";
-import { LandingPageComponent } from './landing-page/landing-page.component';
-import { SingleFaceSnapComponent } from './single-face-snap/single-face-snap.component';
-import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import { NewFaceSnapComponent } from './new-face-snap/new-face-snap.component';
-import {HttpClientModule} from "@angular/common/http";
-import {httpInterceptorProviders} from "./interceptors";
+import {CoreModule} from "./core/core.module";
+import {FaceSnapsModule} from "./face-snaps/face-snaps.module";
+import {LandingPageModule} from "./landing-page/landing-page.module";
 
 @NgModule({
   declarations: [
     AppComponent,
-    FaceSnapComponent,
-    FaceSnapListComponent,
-    HeaderComponent,
-    LandingPageComponent,
-    SingleFaceSnapComponent,
-    NewFaceSnapComponent,
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    LightboxModule,
-    FormsModule,
-    ReactiveFormsModule,
-    HttpClientModule
+    CoreModule,
+    FaceSnapsModule,
+    LandingPageModule
   ],
-  providers: [
-    {provide: LOCALE_ID, useValue: 'fr-FR'},
-    httpInterceptorProviders
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule {
-  constructor() {
-    registerLocaleData(fr.default);
-  }
-}
+export class AppModule {}
