+++ src/app/core/core.module.ts
@@ -1,0 +1,30 @@
+import {LOCALE_ID, NgModule} from '@angular/core';
+import {CommonModule, registerLocaleData} from '@angular/common';
+import {httpInterceptorProviders} from "./interceptors";
+import * as fr from '@angular/common/locales/fr';
+import {HeaderComponent} from "./components/header/header.component";
+import {RouterModule} from "@angular/router";
+import {HttpClientModule} from "@angular/common/http";
+
+@NgModule({
+  declarations: [
+    HeaderComponent
+  ],
+  imports: [
+    CommonModule,
+    RouterModule,
+    HttpClientModule
+  ],
+  exports: [
+    HeaderComponent
+  ],
+  providers: [
+    {provide: LOCALE_ID, useValue: 'fr-FR'},
+    httpInterceptorProviders
+  ]
+})
+export class CoreModule {
+  constructor() {
+    registerLocaleData(fr.default);
+  }
+}
