+++ src/app/core/models/face-snap.model.ts
