+++ src/app/face-snaps/components/face-snap/face-snap.component.scss
