+++ src/app/face-snaps/components/new-face-snap/new-face-snap.component.scss
