+++ src/app/face-snaps/components/new-face-snap/new-face-snap.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {map, Observable, tap} from "rxjs";
-import {FaceSnapModel} from "../models/face-snap.model";
-import {FaceSnapsService} from "../services/face-snaps.service";
+import {FaceSnapModel} from "../../../core/models/face-snap.model";
+import {FaceSnapsService} from "../../../core/services/face-snaps.service";
 import {Router} from "@angular/router";
 
 @Component({
