+++ src/app/face-snaps/components/single-face-snap/single-face-snap.component.scss
