+++ src/app/face-snaps/components/single-face-snap/single-face-snap.component.ts
@@ -1,32 +1,31 @@
-import {Component, Input, OnDestroy, OnInit} from '@angular/core';
-import {FaceSnapModel} from "../models/face-snap.model";
-import {FaceSnapsService} from "../services/face-snaps.service";
+import {Component, OnDestroy, OnInit} from '@angular/core';
+import {FaceSnapModel} from "../../../core/models/face-snap.model";
+import {FaceSnapsService} from "../../../core/services/face-snaps.service";
+import {ActivatedRoute, Router} from "@angular/router";
 import {Lightbox, LightboxConfig} from "ngx-lightbox";
-import {Router} from "@angular/router";
 import {Observable, Subject, takeUntil} from "rxjs";
 
 @Component({
-  selector: 'app-face-snap',
-  templateUrl: './face-snap.component.html',
-  styleUrls: ['./face-snap.component.scss']
+  selector: 'app-single-face-snap',
+  templateUrl: './single-face-snap.component.html',
+  styleUrls: ['./single-face-snap.component.scss']
 })
-export class FaceSnapComponent implements OnInit, OnDestroy {
-  @Input() faceSnap!: FaceSnapModel;
+export class SingleFaceSnapComponent implements OnInit, OnDestroy {
   faceSnap$!: Observable<FaceSnapModel>
   private unsubscribe$ = new Subject<void>();
 
   constructor(
     private _faceSnapsService: FaceSnapsService,
+    private _activatedRoute: ActivatedRoute,
+    private _router: Router,
     private _lightbox: Lightbox,
-    private _lightboxOption: LightboxConfig,
-    private _router: Router
+    private _lightboxOption: LightboxConfig
   ) {
-    _lightboxOption.positionFromTop = 200;
-    _lightboxOption.fadeDuration = 0.3;
+    this._lightboxOption.positionFromTop = 200;
   }
 
   ngOnInit() {
-    const snapId = this.faceSnap.id;
+    const snapId = +this._activatedRoute.snapshot.params['id']; // Add '+' at the beginning of the expression cast a string of numbers in number.
     this.faceSnap$ = this._faceSnapsService.getFaceSnapById(snapId);
   }
 
@@ -45,10 +44,6 @@ export class FaceSnapComponent implements OnInit, OnDestroy {
       );
   }
 
-  onViewFaceSnap() {
-    this._router.navigateByUrl(`facesnaps/${this.faceSnap.id}`);
-  }
-
   onOpenImage(imageUrl: string, description: string): void {
     console.log('Image clicked ' + imageUrl + ' ' + description);
     const album = [{
@@ -57,6 +52,10 @@ export class FaceSnapComponent implements OnInit, OnDestroy {
       thumb: '' // miniature (optionnelle)
     }];
     this._lightbox.open(album, 0);
+  }
+
+  onBack() {
+    this._router.navigateByUrl('facesnaps');
   }
 
   ngOnDestroy() {
