+++ src/app/landing-page/components/landing-page/landing-page.component.scss
