+++ src/app/landing-page/landing-page.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {LandingPageComponent} from "./components/landing-page/landing-page.component";
+import {FormsModule} from "@angular/forms";
+
+@NgModule({
+  declarations: [
+    LandingPageComponent
+  ],
+  imports: [
+    CommonModule,
+    FormsModule
+  ],
+  exports: [
+    LandingPageComponent
+  ]
+})
+export class LandingPageModule { }
